requirements.txt
requests==2.25.1
alive_progress==1.6.2
Pillow==8.2.0
PyQt5==5.15.4
